Add missing tests
See https://github.com/instedd/surveda/pull/1901#discussion_r656376465
