+++ test/lib/runtime/panel_survey_test.exs
@@ -2,8 +2,8 @@ defmodule Ask.Runtime.PanelSurveyTest do
   use Ask.ModelCase
   use Ask.TestHelpers
   use Ask.MockTime
-  alias Ask.Runtime.{PanelSurvey, RespondentGroupAction}
-  alias Ask.{Survey, Repo, TestChannel, Respondent}
+  alias Ask.Runtime.PanelSurvey
+  alias Ask.{Survey, Repo, Respondent}
 
   describe "new_occurrence/1" do
     test "creates a new ready occurrence" do
@@ -201,30 +201,6 @@ defmodule Ask.Runtime.PanelSurveyTest do
 
   defp clean_dates(schedule) do
     schedule |> Map.put(:start_date, nil) |> Map.put(:end_date, nil)
-  end
-
-  defp completed_panel_survey_with_respondents() do
-    panel_survey = panel_survey_with_occurrence()
-    latest_occurrence = Ask.PanelSurvey.latest_occurrence(panel_survey)
-
-    insert_respondents = fn mode, phone_numbers ->
-      channel = TestChannel.new()
-      channel = insert(:channel, settings: channel |> TestChannel.settings(), type: mode)
-      insert_respondents(latest_occurrence, channel, mode, phone_numbers)
-    end
-
-    insert_respondents.("sms", ["1", "2", "3"])
-    insert_respondents.("ivr", ["3", "4"])
-    terminate_survey(latest_occurrence)
-
-    # Reload the panel survey. One of its surveys has changed, so it's outdated
-    Repo.get!(Ask.PanelSurvey, panel_survey.id)
-  end
-
-  defp insert_respondents(survey, channel, mode, phone_numbers) do
-    phone_numbers = RespondentGroupAction.loaded_phone_numbers(phone_numbers)
-    group = RespondentGroupAction.create(UUID.uuid4(), phone_numbers, survey)
-    RespondentGroupAction.update_channels(group.id, [%{"id" => channel.id, "mode" => mode}])
   end
 
   defp set_one_respondent_disposition(survey, disposition) do
