+++ test/support/test_helpers.ex
@@ -2,8 +2,8 @@ defmodule Ask.TestHelpers do
   defmacro __using__(_) do
     quote do
       use Ask.DummySteps
-      alias Ask.Runtime.{Broker, Flow}
-      alias Ask.{PanelSurvey, Repo, Respondent, Survey}
+      alias Ask.Runtime.{Broker, Flow, RespondentGroupAction}
+      alias Ask.{PanelSurvey, Repo, Respondent, Survey, TestChannel}
 
       @foo_string "foo"
       @bar_string "bar"
@@ -166,6 +166,30 @@ defmodule Ask.TestHelpers do
       defp panel_survey_with_last_occurrence_terminated() do
         panel_survey_with_occurrence()
         |> complete_last_occurrence_of_panel_survey()
+      end
+
+      defp completed_panel_survey_with_respondents() do
+        panel_survey = panel_survey_with_occurrence()
+        latest_occurrence = Ask.PanelSurvey.latest_occurrence(panel_survey)
+
+        insert_respondents = fn mode, phone_numbers ->
+          channel = TestChannel.new()
+          channel = insert(:channel, settings: channel |> TestChannel.settings(), type: mode)
+          insert_respondents(latest_occurrence, channel, mode, phone_numbers)
+        end
+
+        insert_respondents.("sms", ["1", "2", "3"])
+        insert_respondents.("ivr", ["3", "4"])
+        terminate_survey(latest_occurrence)
+
+        # Reload the panel survey. One of its surveys has changed, so it's outdated
+        Repo.get!(Ask.PanelSurvey, panel_survey.id)
+      end
+
+      defp insert_respondents(survey, channel, mode, phone_numbers) do
+        phone_numbers = RespondentGroupAction.loaded_phone_numbers(phone_numbers)
+        group = RespondentGroupAction.create(UUID.uuid4(), phone_numbers, survey)
+        RespondentGroupAction.update_channels(group.id, [%{"id" => channel.id, "mode" => mode}])
       end
     end
   end
